a few more split
