+++ static/app/gettingStartedDocs/python/python/agentMonitoring.tsx
@@ -5,7 +5,8 @@ import type {
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {t, tct} from 'sentry/locale';
-import {getPythonInstallCodeBlock} from 'sentry/utils/gettingStartedDocs/python';
+
+import {getPythonInstallCodeBlock} from './utils';
 
 export const agentMonitoring: OnboardingConfig = {
   install: params => {
