+++ static/app/gettingStartedDocs/python/tryton.tsx
@@ -8,12 +8,9 @@ import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {agentMonitoring} from 'sentry/gettingStartedDocs/python/python/agentMonitoring';
 import {crashReport} from 'sentry/gettingStartedDocs/python/python/crashReport';
 import {mcp} from 'sentry/gettingStartedDocs/python/python/mcp';
+import {alternativeProfiling} from 'sentry/gettingStartedDocs/python/python/profiling';
+import {getPythonInstallCodeBlock} from 'sentry/gettingStartedDocs/python/python/utils';
 import {t, tct} from 'sentry/locale';
-import {
-  alternativeProfilingConfiguration,
-  getPythonInstallCodeBlock,
-  getPythonLogsOnboarding,
-} from 'sentry/utils/gettingStartedDocs/python';
 
 type Params = DocsParams;
 
@@ -137,14 +134,12 @@ const onboarding: OnboardingConfig = {
             },
           ],
         },
-        alternativeProfilingConfiguration(params),
+        alternativeProfiling(params),
       ],
     },
   ],
   verify: () => [],
 };
-
-const logsOnboarding = getPythonLogsOnboarding();
 
 const profilingOnboarding: OnboardingConfig = {
   install: onboarding.install,
@@ -170,7 +165,7 @@ const docs: Docs = {
   crashReportOnboarding: crashReport,
   agentMonitoringOnboarding: agentMonitoring,
   mcpOnboarding: mcp,
-  logsOnboarding,
+  logsOnboarding: logs(),
 };
 
 export default docs;
